+++ CHANGELOG.md
@@ -8,6 +8,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- Migration: Fixed splitSnapProps bug not properly handling additional properties
+
 ## [1.30.1] - 2024-12-17
 
 ### Fixed